refactor: 코드 리팩터링 및 파일 분할
