+++ src/js/stores/index.js
@@ -1,0 +1,8 @@
+export const menuStore = {
+  getLocalStorage() {
+    return JSON.parse(localStorage.getItem("menu"));
+  },
+  setLocalStorage(value) {
+    localStorage.setItem("menu", JSON.stringify(value));
+  },
+};
